Add city search functionality
This is a pretty big commit.

This makes it so that the user can search for a city, select the
city from a list, and have that city populated in the city list with
the weather type, city name, and current temp.

This includes work with both the DOM, with the different fetch
modules, and with the template for the City which includes the
city data and the weather data.