Fix weather codes
The weather codes before had names that had to do with the icon
names, but this did not look great for use in other places. I created
a sublayer in the weathercodes object so that the icon or the name
could be selected.

I also fixed some of the typography to prepare for getting it to look
good for mobile use.