Bug fix
The code checks to see if the weather page city is undefined, and if
the city is deleted, the status was not changed to undefined. This
does something potentially naughty by defining the city as undefined
if the city is removed from the city list. It seems like the most
elegent solution. Perhaps ill go back and fix it later.