a11y changes
This makes some of the divs into real buttons. When they were converted
into buttons, the styles needed tweaking too.

This allowed the user to tab through the options which is nice. This did
cause the tab order for the cities to be wonky, so I needed to legitly
sort the cities list so that they are logically set out and so that tab
order actually works.

I also made the font larger when in mobile view.