Change icon set
In the near future, I want to include icons and maybe pictures that
reflect if it is night or day, so I switched to an icon set that has day
and night weather icons.

The icons also needed a special folder structure, so I had to tweak
the folders a bit.