Add data flip method
The way that the data is currently, it has an array for time and an
array for the output. The way that plots work is that they require
that the time and data are bundled in the same object, so this makes
the time and the output into one object that is then combined into
an array. This is preliminary work for getting plots up and going.