Fix precipitation chart
It was not great how it was scaled. There was no clear way to know what
heavy, light, moderate rain was. I found an estimate about what these
categories are in in/hr, so I just used that scale for everything.

This also eliminates the need for the "no rain in next three days" text.